+++ data/nyanko.txt
@@ -1,3 +1,3 @@
 ToDo|0|hi
 Deadline|1|hw|2020-09-22 1223
-Deadline|0|hw|2323-01-01 1111
+Event|0|lol|2012-09-09 0909|2012-09-09 2222
