+++ src/main/java/nyanko/command/ByeCommand.java
@@ -1,3 +1,9 @@
+package nyanko.command;
+
+import nyanko.storage.Storage;
+import nyanko.task.TaskList;
+import nyanko.ui.Ui;
+
 public class ByeCommand extends Command {
     @Override
     public void execute(TaskList tasks, Ui ui, Storage storage) {
