+++ src/main/java/nyanko/task/InvalidTaskFormatException.java
@@ -1,3 +1,5 @@
+package nyanko.task;
+
 public class InvalidTaskFormatException extends Exception {
     public InvalidTaskFormatException(String message) {
         super(message);
