+++ src/main/java/nyanko/task/Task.java
@@ -1,3 +1,5 @@
+package nyanko.task;
+
 public class Task {
     protected String description;
     protected boolean isDone;
