+++ src/main/java/nyanko/task/TaskList.java
@@ -1,3 +1,5 @@
+package nyanko.task;
+
 import java.util.ArrayList;
 
 public class TaskList {
