+++ src/main/java/nyanko/ui/Ui.java
@@ -1,3 +1,5 @@
+package nyanko.ui;
+
 import java.util.Scanner;
 
 public class Ui {
